+++ mods/src/misc/misc.cpp
@@ -49,6 +49,7 @@
 #include <symbols/ReedsFeature.h>
 #include <symbols/Feature.h>
 #include <symbols/Biome.h>
+#include <symbols/CreatorMode.h>
 
 #include <mods/init/init.h>
 #include <mods/feature/feature.h>
@@ -112,6 +113,14 @@ static bool FurnaceScreen_handleAddItem_injection(FurnaceScreen_handleAddItem_t 
 }
 static bool FurnaceScreen_moveOver_FillingContainer_removeResource_two_injection(FillingContainer *self, const ItemInstance &item) {
     return self->removeResource_one(item, true) == 0;
+}
+
+// Fix Creative Mode Mining Delay
+static void CreatorMode_stopDestroyBlock_injection(CreatorMode_stopDestroyBlock_t original, CreatorMode *self) {
+    // Call Original Method
+    original(self);
+    // Reset Delay
+    self->destroy_delay = 0;
 }
 
 // Get Real Selected Slot
@@ -533,9 +542,8 @@ void init_misc() {
     }
 
     // Disable Creative Mode Mining Delay
-    if (feature_has("Disable Creative Mode Mining Delay", server_disabled)) {
-        unsigned char nop_patch[4] = {0x00, 0xf0, 0x20, 0xe3}; // "nop"
-        patch((void *) 0x19fa0, nop_patch);
+    if (feature_has("Fix Blocks Not Breaking Instantly In Creative Mode", server_enabled)) {
+        overwrite_calls(CreatorMode_stopDestroyBlock, CreatorMode_stopDestroyBlock_injection);
     }
 
     // Generate Caves